Refactor the API
