+++ cruft/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Contains the core logic behind all cruft commands."""
