+++ tests/test_api.py
@@ -4,8 +4,7 @@ import os
 import pytest
 from examples import verify_and_test_examples
 from git import Repo
-from git.exc import GitCommandError
-from hypothesis_auto import auto_pytest, auto_pytest_magic
+from hypothesis_auto import auto_pytest
 
 from cruft import api, exceptions
